# docs/implementation/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/anatomy/structures/{id}": {
      "get": {
        "operationId": "PublicController_structure",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Public"
        ]
      }
    },
...
    "/api/v1/assets/manifests/{id}": {
      "get": {
        "operationId": "PublicController_manifest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Public"
        ]
      }
    },
    "/api/v1/assets/versions/{id}/access": {
      "post": {
        "operationId": "PublicController_access",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Public"
        ]
      }
    },
...
    "/api/v1/articles/{slug}": {
      "get": {
        "operationId": "PublicController_article",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Public"
        ]
      }
    },
...
    "/api/v1/notes/{id}": {
      "patch": {
        "operationId": "PrivateController_updateNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      },
      "delete": {
        "operationId": "PrivateController_deleteNote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
    "/api/v1/lessons": {
      "get": {
        "operationId": "PrivateController_lessons",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      },
      "post": {
        "operationId": "PrivateController_createLesson",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
    "/api/v1/lessons/{id}": {
      "get": {
        "operationId": "PrivateController_lesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      },
      "patch": {
        "operationId": "PrivateController_updateLesson",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
    "/api/v1/lessons/{id}/presentation": {
      "get": {
        "operationId": "PrivateController_presentation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
    "/api/v1/lessons/{id}/scenes": {
      "post": {
        "operationId": "PrivateController_addScene",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
    "/api/v1/scenes/{id}": {
      "put": {
        "operationId": "PrivateController_updateScene",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
...
  },
  "info": {
    "title": "HumanScope REST API",
    "description": "Foundation route inventory. Body validation is defined by Zod contracts; response schemas and full OpenAPI coverage remain a release gate.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hs_session"
      }
    },
...
  }
}